Precoded app version.
