+++ xboinc/simulation_io/__init__.py
@@ -4,5 +4,6 @@
 # ######################################### #
 
 from .default_tracker import get_default_tracker
-from .output import SimState, SimVersion
+from .output import SimState
 from .input import SimConfig
+from .version import SimVersion, app_version, app_version_int
